+++ site/src/components/SettingsLayout/Section.tsx
@@ -1,7 +1,7 @@
 import { makeStyles } from "@mui/styles"
 import Typography from "@mui/material/Typography"
 import { FC, ReactNode, PropsWithChildren } from "react"
-import { SectionAction } from "../SectionAction/SectionAction"
+import { SectionAction } from "./SectionAction"
 
 type SectionLayout = "fixed" | "fluid"
